Move timeit to stdx
